Rename package name to better reflect it's nature
