+++ eureka-client/src/main/java/com/netflix/eureka2/client/EurekaClientBuilder.java
@@ -5,8 +5,8 @@ import com.netflix.eureka2.client.channel.ClientChannelFactoryImpl;
 import com.netflix.eureka2.client.metric.EurekaClientMetricFactory;
 import com.netflix.eureka2.client.registration.RegistrationHandler;
 import com.netflix.eureka2.client.registration.RegistrationHandlerImpl;
-import com.netflix.eureka2.client.registry.InterestHandler;
-import com.netflix.eureka2.client.registry.InterestHandlerImpl;
+import com.netflix.eureka2.client.interest.InterestHandler;
+import com.netflix.eureka2.client.interest.InterestHandlerImpl;
 import com.netflix.eureka2.client.resolver.ServerResolver;
 import com.netflix.eureka2.client.transport.TransportClients;
 import com.netflix.eureka2.config.BasicEurekaRegistryConfig;
