+++ eureka-client/src/main/java/com/netflix/eureka2/client/interest/swap/RegistrySwapOperator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.eureka2.client.registry.swap;
+package com.netflix.eureka2.client.interest.swap;
 
 import com.netflix.eureka2.interests.ChangeNotification;
 import com.netflix.eureka2.registry.EurekaRegistry;
