+++ eureka-client/src/test/java/com/netflix/eureka2/client/EurekaClientTest.java
@@ -3,8 +3,8 @@ package com.netflix.eureka2.client;
 import com.netflix.eureka2.client.channel.ClientChannelFactory;
 import com.netflix.eureka2.client.channel.ClientChannelFactoryImpl;
 import com.netflix.eureka2.client.metric.EurekaClientMetricFactory;
-import com.netflix.eureka2.client.registry.InterestHandler;
-import com.netflix.eureka2.client.registry.InterestHandlerImpl;
+import com.netflix.eureka2.client.interest.InterestHandler;
+import com.netflix.eureka2.client.interest.InterestHandlerImpl;
 import com.netflix.eureka2.config.BasicEurekaRegistryConfig;
 import com.netflix.eureka2.interests.ChangeNotification;
 import com.netflix.eureka2.interests.Interest;
